requirements.txt
Flask==1.1.2
dynaconf==3.1.2
Flask-SQLAlchemy==2.4.4
sentry-sdk
marshmallow==3.7.0
gunicorn==20.0.4
python-dotenv==0.14.0
requests==2.25.0
Flask-RESTful==0.3.8
pycpfcnpj==1.5.1
psycopg2==2.8.5
SQLAlchemy-serializer==1.3.4.4
arrow==0.17.0
grpcio==1.41.1
grpcio-tools==1.41.1

